Refactor: public 키워드 제거
